if you try to be defensive then do it right
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/compress/trunk@1548924 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/compress/compressors/snappy/SnappyCompressorInputStream.java
@@ -287,7 +287,7 @@ public class SnappyCompressorInputStream extends CompressorInputStream {
             length = readOneByte();
             length |= (readOneByte() << 8);
             length |= (readOneByte() << 16);
-            length |= (readOneByte() << 24);
+            length |= (((long) readOneByte()) << 24);
             break;
         default:
             length = b >> 2;
